Refactor to generic IdP discovery with Azure + Okta support
Rename `discover azure tenant` → `discover idp`. Now detects multiple
identity providers for a domain:

Azure/Entra ID: OpenID config, UserRealm, GetCredentialType, M365 services
Okta: DNS CNAME detection, OpenID config on SSO subdomains, Azure
  UserRealm federation URL analysis

Signal returns a list of discovered providers with provider-specific
detail blocks, supporting domains that use multiple IdPs.
